Move consts to utils to fix build error
