+++ web/lib/types/common.ts
@@ -1,5 +1,5 @@
 import { DocumentData, QueryDocumentSnapshot, Timestamp } from "firebase/firestore";
-import { BABIES_TAB, CAREGIVERS_TAB } from "pages/consts";
+import { BABIES_TAB, CAREGIVERS_TAB } from "@lib/utils/consts";
 
 export interface User {
   name: string;
